branch 변경

+++ src/main/kotlin/com/example/cakezip/domain/Review.kt
@@ -1,10 +1,6 @@
 package com.example.cakezip.domain
 
-import com.example.cakezip.domain.BaseEntity
 import com.example.cakezip.domain.cake.Cake
-import com.example.cakezip.domain.member.Customer
-import com.example.cakezip.domain.shop.Shop
-import lombok.*
 import javax.persistence.*
 
 @Entity
@@ -15,7 +11,7 @@ class Review(
 
     val reviewTitle: String,
     val reviewContent: String,
-    val reviewScore: Float,
+    val reviewScore: Int,
 
     @OneToOne
     @JoinColumn(name = "cake_id")
